Do tag counting with org-map-trees
Instead of just re-search-forward this allows us to easily exclude
archived trees.

This is slower though. About 7 times slower for one of my 500K
repository and 6 times for the other. Hard to tell so much about the
scaling behaviour but in these cases it’s still just fractions of a
second for loading the counts on, so this will probably not be a big
problem.
Otherwise it could be made switchable and the old re-search-forward
approach used if ignoring archives etc. is not wanted.