helm-tags: New method for avoiding matches of count
Using helm-tags-match-part is not optimal since it assumes that the
part to match is a substring of the displayed string when doing
match-highlighting. This didn’t work so well when I attempted to
construct the match-part as the concatenation of tag and for
display strings like these:
tag    count info

The current method is to not store count at all in the display string,
but store spaces and display it with the display text-property there.
This works really well and ensures we will match on tag and info.

In the future this could easily be customized with some
options (match-count? match-info?), and different methods chosen in
orgqda-helm-tags--format-list-item.