Fix extraction from inlinetasks before first headline
Previously, org-get-heading would return nil when we were before the
first real headline. But we know that we are on a headline or
inlinetask so org-get-heading should work.

This may be a bug that should be reported to org though. What is the
purpose of org-get-heading for inlinetasks? Perhaps I’m doing it wrong
here anyway.