Completion: Move advice for getting annotated tag list
The problem with the earlier approach was that org-set-tags-command
tries to read and merge both a tag list from the current
buffer (local) and a global tag list (from agenda files). Previously
we advised the global tag list
fetcher (org-global-tags-completion-table), but as we add text
properties to the tags, this wasn’t merged cleanly with local tags.

The current approach of just overriding the fetching of buffer-local
tags (org-get-buffer-tags) avoids this problem, and also avoids doing
unnecessary double work in the current buffer (the orgqda tag
collection is usually done in current buffer and others).